+++ bee-ledger/src/workers/consensus/worker_del.rs
@@ -2,62 +2,45 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
-    types::{Balance, CreatedOutput, LedgerIndex, Migration, Receipt, TreasuryOutput},
+    types::{CreatedOutput, LedgerIndex, Migration, Receipt, TreasuryOutput},
     workers::{
         consensus::{metadata::WhiteFlagMetadata, state::validate_ledger_state, white_flag},
         error::Error,
         event::{MilestoneConfirmed, OutputConsumed, OutputCreated},
-        pruning::{condition::should_prune, config::PruningConfig, prune},
+        pruning::{self, condition::should_prune, config::PruningConfig},
         snapshot::{condition::should_snapshot, config::SnapshotConfig, worker::SnapshotWorker},
         storage::{self, StorageBackend},
     },
 };
 
+use bee_ledger_types::types::ConflictReason;
 use bee_message::{
-    address::Address,
     milestone::MilestoneIndex,
     output::{Output, OutputId},
     payload::{milestone::MilestoneId, receipt::ReceiptPayload, transaction::TransactionId, Payload},
     MessageId,
 };
 use bee_runtime::{event::Bus, node::Node, shutdown_stream::ShutdownStream, worker::Worker};
-use bee_tangle::{ConflictReason, MsTangle, TangleWorker};
+use bee_tangle::{MsTangle, TangleWorker};
 
 use async_trait::async_trait;
 
-use futures::{channel::oneshot, stream::StreamExt};
+use futures::stream::StreamExt;
 use log::{error, info, warn};
 use tokio::sync::mpsc;
 use tokio_stream::wrappers::UnboundedReceiverStream;
 
 use std::{any::TypeId, convert::TryInto};
 
-pub(crate) const EXTRA_SNAPSHOT_DEPTH: u32 = 5;
-pub(crate) const EXTRA_PRUNING_DEPTH: u32 = 5;
+const SNAPSHOT_DEPTH_MIN: u32 = 50;
 
-/// Commands of the consensus worker.
-#[allow(clippy::type_complexity)]
-pub enum ConsensusWorkerCommand {
-    /// Command to confirm a milestone.
-    ConfirmMilestone(MessageId),
-    /// Command to fetch the balance of an address.
-    FetchBalance(Address, oneshot::Sender<(Result<Option<Balance>, Error>, LedgerIndex)>),
-    /// Command to fetch an output.
-    FetchOutput(
-        OutputId,
-        oneshot::Sender<(Result<Option<CreatedOutput>, Error>, LedgerIndex)>,
-    ),
-    /// Command to fetch the outputs of an address.
-    FetchOutputs(
-        Address,
-        oneshot::Sender<(Result<Option<Vec<OutputId>>, Error>, LedgerIndex)>,
-    ),
-}
+const PRUNING_DEPTH_DELTA: u32 = 50;
+const PRUNING_INTERVAL: u32 = 10;
 
-/// The consensus worker.
+pub struct ConsensusWorkerEvent(pub MessageId);
+
 pub struct ConsensusWorker {
-    /// Communication channel of the consensus worker.
-    pub tx: mpsc::UnboundedSender<ConsensusWorkerCommand>,
+    pub tx: mpsc::UnboundedSender<ConsensusWorkerEvent>,
 }
 
 pub(crate) async fn migration_from_milestone(
@@ -74,13 +57,9 @@ pub(crate) async fn migration_from_milestone(
         match receipt.inner().transaction() {
             Payload::TreasuryTransaction(treasury) => match treasury.output() {
                 Output::Treasury(output) => output.clone(),
-                Output::SignatureLockedDustAllowance(_) | Output::SignatureLockedSingle(_) => {
-                    return Err(Error::UnsupportedOutputKind(treasury.output().kind()));
-                }
+                output => return Err(Error::UnsupportedOutputKind(output.kind())),
             },
-            Payload::Milestone(_) | Payload::Indexation(_) | Payload::Receipt(_) | Payload::Transaction(_) => {
-                return Err(Error::UnsupportedPayloadKind(receipt.inner().transaction().kind()));
-            }
+            payload => return Err(Error::UnsupportedPayloadKind(payload.kind())),
         },
         milestone_id,
     );
@@ -99,10 +78,7 @@ async fn confirm<N: Node>(
 where
     N::Backend: StorageBackend,
 {
-    let message = tangle
-        .get(&message_id)
-        .await
-        .ok_or(Error::MilestoneMessageNotFound(message_id))?;
+    let message = tangle.get(&message_id).await.ok_or(Error::MilestoneMessageNotFound)?;
 
     let milestone = match message.payload() {
         Some(Payload::Milestone(milestone)) => milestone.clone(),
@@ -110,7 +86,7 @@ where
     };
 
     if milestone.essence().index() != MilestoneIndex(**ledger_index + 1) {
-        return Err(Error::NonContiguousMilestones(
+        return Err(Error::NonContiguousMilestone(
             *milestone.essence().index(),
             **ledger_index,
         ));
@@ -118,11 +94,14 @@ where
 
     let mut metadata = WhiteFlagMetadata::new(milestone.essence().index());
 
-    white_flag(tangle, storage, message.parents(), &mut metadata).await?;
+    let parents = message.parents().iter().copied().collect();
+
+    drop(message);
+
+    white_flag::<N::Backend>(tangle, storage, parents, &mut metadata).await?;
 
     if !metadata.merkle_proof.eq(&milestone.essence().merkle_proof()) {
         return Err(Error::MerkleProofMismatch(
-            milestone.essence().index(),
             hex::encode(metadata.merkle_proof),
             hex::encode(milestone.essence().merkle_proof()),
         ));
@@ -149,10 +128,7 @@ where
         }
 
         if receipt.migrated_at() < *receipt_migrated_at {
-            return Err(Error::DecreasingReceiptMigratedAtIndex(
-                receipt.migrated_at(),
-                *receipt_migrated_at,
-            ));
+            return Err(Error::DecreasingReceiptMigratedAtIndex);
         } else {
             *receipt_migrated_at = receipt.migrated_at();
         }
@@ -227,7 +203,7 @@ where
     );
 
     bus.dispatch(MilestoneConfirmed {
-        message_id,
+        id: message_id,
         index: milestone.essence().index(),
         timestamp: milestone.essence().timestamp(),
         referenced_messages: metadata.referenced_messages,
@@ -265,125 +241,82 @@ where
     }
 
     async fn start(node: &mut N, config: Self::Config) -> Result<Self, Self::Error> {
-        let (snapshot_config, pruning_config) = config;
+        let (mut snapshot_config, mut pruning_config) = config;
+
         let (tx, rx) = mpsc::unbounded_channel();
+
         let tangle = node.resource::<MsTangle<N::Backend>>();
         let storage = node.storage();
         let bus = node.bus();
 
         validate_ledger_state(&*storage).await?;
 
-        let bmd = tangle.config().below_max_depth();
-
-        let snapshot_depth_min = bmd + EXTRA_SNAPSHOT_DEPTH;
-        let snapshot_depth = if snapshot_config.depth() < snapshot_depth_min {
+        if snapshot_config.depth() < SNAPSHOT_DEPTH_MIN {
             warn!(
                 "Configuration value for \"depth\" is too low ({}), value changed to {}.",
                 snapshot_config.depth(),
-                snapshot_depth_min
+                SNAPSHOT_DEPTH_MIN
             );
-            snapshot_depth_min
-        } else {
-            snapshot_config.depth()
-        };
+            snapshot_config.override_depth(SNAPSHOT_DEPTH_MIN)
+        }
 
-        let snapshot_pruning_delta = bmd + EXTRA_PRUNING_DEPTH;
-
-        let pruning_depth_min = snapshot_depth + snapshot_pruning_delta;
-        let pruning_depth = if pruning_config.delay() < pruning_depth_min {
+        let pruning_delay_min = snapshot_config.depth() + PRUNING_DEPTH_DELTA;
+        if pruning_config.delay() < pruning_delay_min {
             warn!(
                 "Configuration value for \"delay\" is too low ({}), value changed to {}.",
                 pruning_config.delay(),
-                pruning_depth_min
+                pruning_delay_min
             );
-            pruning_depth_min
-        } else {
-            pruning_config.delay()
-        };
+            pruning_config.override_delay(pruning_delay_min);
+        }
 
-        // Unwrap is fine because ledger index was already in storage or just added by the snapshot worker.
-        let mut ledger_index = storage::fetch_ledger_index(&*storage).await?.unwrap();
+        // Unwrap is fine because we just inserted the ledger index.
+        // TODO unwrap
+        let mut ledger_index = storage::fetch_ledger_index(&*storage).await.unwrap().unwrap();
         let mut receipt_migrated_at = MilestoneIndex(0);
+        let below_max_depth = tangle.config().below_max_depth();
 
         node.spawn::<Self, _, _>(|shutdown| async move {
             info!("Running.");
 
             let mut receiver = ShutdownStream::new(shutdown, UnboundedReceiverStream::new(rx));
 
-            while let Some(event) = receiver.next().await {
-                match event {
-                    ConsensusWorkerCommand::ConfirmMilestone(message_id) => {
-                        if let Err(e) = confirm::<N>(
-                            &tangle,
-                            &storage,
-                            &bus,
-                            message_id,
-                            &mut ledger_index,
-                            &mut receipt_migrated_at,
-                        )
-                        .await
-                        {
-                            error!("Confirmation error on {}: {}.", message_id, e);
-                            panic!("Aborting due to unexpected ledger error.");
-                        }
+            while let Some(ConsensusWorkerEvent(message_id)) = receiver.next().await {
+                if let Err(e) = confirm::<N>(
+                    &tangle,
+                    &storage,
+                    &bus,
+                    message_id,
+                    &mut ledger_index,
+                    &mut receipt_migrated_at,
+                )
+                .await
+                {
+                    error!("Confirmation error on {}: {}.", message_id, e);
+                    panic!("Aborting due to unexpected ledger error.");
+                }
 
-                        if !tangle.is_confirmed() {
-                            continue;
-                        }
+                if !tangle.is_confirmed() {
+                    continue;
+                }
 
-                        match should_snapshot(&tangle, ledger_index, pruning_depth_min, &snapshot_config) {
-                            Ok(()) => {
-                                // TODO
-                                // if let Err(e) = snapshot(snapshot_config.path(), event.index - snapshot_depth) {
-                                //     error!("Failed to create snapshot: {:?}.", e);
-                                // }
-                            }
-                            Err(reason) => {
-                                info!("Snapshot skipped: {:?}", reason);
-                            }
-                        }
-
-                        match should_prune(&tangle, ledger_index, pruning_depth, &pruning_config) {
-                            Ok(target_index) => {
-                                if let Err(e) = prune::prune(
-                                    &tangle,
-                                    &storage,
-                                    &bus,
-                                    target_index,
-                                    snapshot_pruning_delta,
-                                    &pruning_config,
-                                )
+                if let Ok(_snapshot_index) = should_snapshot(&tangle, ledger_index, &snapshot_config) {
+                    // if let Err(e) = snapshot(snapshot_config.path(), snapshot_index) {
+                    //     error!("Failed to create snapshot: {:?}.", e);
+                    // }
+                }
+                match should_prune(&tangle, ledger_index, &pruning_config) {
+                    Ok(target_index) => {
+                        if let Err(e) =
+                            pruning::prune(&tangle, &storage, &bus, target_index, &pruning_config, below_max_depth)
                                 .await
-                                {
-                                    error!("Pruning failed: {:?}.", e);
-                                }
-                            }
-                            Err(reason) => {
-                                info!("Pruning skipped: {:?}", reason);
-                            }
-                        }
-                    }
-                    ConsensusWorkerCommand::FetchBalance(address, sender) => {
-                        if let Err(e) = sender.send((storage::fetch_balance(&*storage, &address).await, ledger_index)) {
-                            error!("Error while sending balance: {:?}", e);
-                        }
-                    }
-                    ConsensusWorkerCommand::FetchOutput(output_id, sender) => {
-                        if let Err(e) = sender.send((storage::fetch_output(&*storage, &output_id).await, ledger_index))
                         {
-                            error!("Error while sending output: {:?}", e);
+                            error!("Failed to prune database: {:?}.", e);
                         }
                     }
-                    ConsensusWorkerCommand::FetchOutputs(address, sender) => match address {
-                        Address::Ed25519(address) => {
-                            if let Err(e) = sender.send((
-                                storage::fetch_outputs_for_ed25519_address(&*storage, &address).await,
-                                ledger_index,
-                            )) {
-                                error!("Error while sending output: {:?}", e);
-                            }
-                        }
-                    },
+                    Err(reason) => {
+                        info!("Pruning skipped. Reason: {:?}", reason);
+                    }
                 }
             }