+++ bee-ledger/src/workers/pruning/condition.rs
@@ -1,77 +1,47 @@
-// Copyright 2020-2021 IOTA Stiftung
+// Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::workers::{
-    pruning::{
-        config::PruningConfig,
-        constants::{PRUNING_THRESHOLD, SOLID_ENTRY_POINT_THRESHOLD_PAST},
-    },
-    snapshot::config::SnapshotConfig,
-};
+use super::config::PruningConfig;
+
+use crate::types::LedgerIndex;
 
 use bee_message::milestone::MilestoneIndex;
-
 use bee_tangle::{storage::StorageBackend, MsTangle};
 
-pub(crate) fn should_snapshot<B: StorageBackend>(
-    tangle: &MsTangle<B>,
-    index: MilestoneIndex,
-    depth: u32,
-    config: &SnapshotConfig,
-) -> bool {
-    let solid_index = *index;
-    let snapshot_index = *tangle.get_snapshot_index();
-    let pruning_index = *tangle.get_pruning_index();
-    let snapshot_interval = if tangle.is_synced() {
-        config.interval_synced()
-    } else {
-        config.interval_unsynced()
-    };
+type PruningTargetIndex = MilestoneIndex;
 
-    if (solid_index < depth + snapshot_interval)
-        || (solid_index - depth) < pruning_index + 1 + SOLID_ENTRY_POINT_THRESHOLD_PAST
-    {
-        // Not enough history to calculate solid entry points.
-        return false;
-    }
+/// Reasons for not-pruning.
+#[derive(Debug)]
+pub enum SkipReason {
+    /// Pruning is disabled in the config.
+    Disabled,
 
-    solid_index - (depth + snapshot_interval) >= snapshot_index
+    /// There is not enough data yet to be pruned.
+    BelowThreshold { reached_in: u32 },
 }
 
 pub(crate) fn should_prune<B: StorageBackend>(
     tangle: &MsTangle<B>,
-    mut index: MilestoneIndex,
-    delay: u32,
-    config: &PruningConfig,
-) -> bool {
-    if !config.enabled() {
-        return false;
+    ledger_index: LedgerIndex,
+    pruning_depth: u32,
+    pruning_config: &PruningConfig,
+) -> Result<PruningTargetIndex, SkipReason> {
+    let pruning_index = *tangle.get_pruning_index();
+    let start_index_min = || pruning_index + pruning_depth + 1;
+
+    if pruning_config.disabled() {
+        // Disabled in the config.
+        Err(SkipReason::Disabled)
+    } else if *ledger_index < start_index_min() {
+        // Not enough history yet.
+        Err(SkipReason::BelowThreshold {
+            reached_in: start_index_min() - *ledger_index,
+        })
+    } else {
+        // NB:
+        // We made sure that `confirmed_milestone_index` >= pruning_depth.
+        let pruning_target_index = *ledger_index - pruning_depth;
+
+        Ok(pruning_target_index.into())
     }
-
-    if *index <= delay {
-        return false;
-    }
-
-    // Pruning happens after creating the snapshot so the metadata should provide the latest index.
-    if *tangle.get_snapshot_index() < SOLID_ENTRY_POINT_THRESHOLD_PAST + PRUNING_THRESHOLD + 1 {
-        return false;
-    }
-
-    let target_index_max =
-        MilestoneIndex(*tangle.get_snapshot_index() - SOLID_ENTRY_POINT_THRESHOLD_PAST - PRUNING_THRESHOLD - 1);
-
-    if index > target_index_max {
-        index = target_index_max;
-    }
-
-    if tangle.get_pruning_index() >= index {
-        return false;
-    }
-
-    // We prune in "PRUNING_THRESHOLD" steps to recalculate the solid_entry_points.
-    if *tangle.get_entry_point_index() + PRUNING_THRESHOLD + 1 > *index {
-        return false;
-    }
-
-    true
 }